Moved classes to a different package
