+++ dialect/src/main/java/ydb/jimmer/client/UuidTransactionCacheOperator.java
@@ -1,4 +1,4 @@
-package ydb.jimmer.cache;
+package ydb.jimmer.client;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
